+++ bin/laravel
@@ -7,7 +7,7 @@ if (file_exists(__DIR__.'/../../../autoload.php')) {
     require __DIR__.'/../vendor/autoload.php';
 }
 
-$app = new Symfony\Component\Console\Application('Laravel Installer', '5.10.0');
+$app = new Symfony\Component\Console\Application('Laravel Installer', '5.11.0');
 $app->add(new Laravel\Installer\Console\NewCommand);
 
 $app->run();
